Reorganize components and pages
