+++ src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import './App.css';
 
-import HomePage from './HomePage';
+import HomePage from './pages/HomePage';
 
 class App extends Component {
   render() {
